Split json and yaml resources
